refactor: #134 - KeyChainManager 프로토콜 구현
